Calculates Significant Metabolic Pathways using a Fisher Exact Test

Calculates Significant Metabolic Pathways using a Fisher Exact Test

FishersPathwayAnalysis(
  Analyte,
  analyte_type = c("mz"),
  polarity = NULL,
  ppm_error = 10,
  max_path_size = 500,
  min_path_size = 5,
  alternative = "greater",
  pathway_all_info = FALSE,
  pval_cutoff = 0.05,
  verbose = TRUE
)

Arguments

Analyte

A list of analytes with 3 elements, namely "mz", "genes" and "metabolites", each is comprised of the corresponding labels, for metabolites, Supported metabolites format including, X stands for upper case of the cooresponding ID in each database: "hmdb:HMDBX", "chebi:X", "pubchem:X","wikidata:X" ,"kegg:X" ,"CAS:X","lipidbank:X","chemspider:X"," LIPIDMAPS:X" Supported gene data format including: "entrez:X", "gene_symbol:X", "uniprot:X", "ensembl:X", "hmdb:HMDBPX" Supported mz format: any string or numeric vector contains the m/z

analyte_type

= "metabolites" or "gene" or "mz", or a vector contains any combinations of them (default = c("mz", "genes")).

polarity

Character string defining the polarity of the MALDI experiment. Inputs must be either 'positive', 'negative' or 'neutral' (default = NULL).

ppm_error

Integer defining the ppm threshold that matched analytes must be between (default = 10).

max_path_size

The max number of metabolites in a specific pathway (default = 500).

min_path_size

The min number of metabolites in a specific pathway (default = 5).

alternative

The hypothesis of the fisher exact test (default = "greater").

pathway_all_info

Whether to included all genes/metabolites screened in the return (default = FALSE).

pval_cutoff

The cut off of raw p value to retain the pathways (default = 0.05).

verbose

Boolean indicating whether to show the message. If TRUE the message will be show, else the message will be suppressed (default = TRUE).

Value

a dataframe with the relevant pathway information

Examples

# FishersPathwayAnalysis(Analyte = mzs, analyte_type = "mz", ppm_error = 3)