Creates a transformation matrix that mirrors an object in 2D along either the x axis or y axis around its center of mass

Creates a transformation matrix that mirrors an object in 2D along either the x axis or y axis around its center of mass

mirror(mirror.x = FALSE, mirror.y = FALSE, center.cur)

Arguments

mirror.x, mirror.y

Logical specifying whether or not an object should be reflected

center.cur

Coordinates of the current center of mass