This the function used to compute the exact fisher test for over-representation based pathway analysis

This the function used to compute the exact fisher test for over-representation based pathway analysis

Usage

pathway_analysis(
  object,
  group.by = NULL,
  polarity = "positive",
  SM.assay = "SPM",
  ST.assay = NULL,
  analyte_type = c("mz", "genes"),
  ...
)

Arguments

object

A SpaMTP Seurat class object contains spatial metabolomics/transcriptomics features or both.

group.by

Character string defining the metadata column that contains the relative groupings. If NULL, will use whole sample as one group (default = NULL).

polarity

The polarity of the MALDI experiment. Inputs must be either NULL, 'positive' or 'negative'. If NULL, pathway analysis will run in neutral mode (default = NULL).

SM.assay

Character string defining the SpaMTP assay that contains m/z values (default = "SPM").

ST.assay

Character string defining the SpaMTP assay that contains gene names (default = NULL).

analyte_type

Vector of character string defining the annotations to use for pathway analysis. Options can be one or multiple of: "metabolites" or "gene" or "mz" (default = c("mz", "gene")).

...

Additional arguments to pass to FisherexactTest. Check FisherexactTest documentation for possible inputs.

Value

A dataframe with the relevant pathway information

Examples

# pathway_analysis(Seurat.Obj, polarity = "positive")