PCA driven pathway analaysis

PCA driven pathway analaysis

Usage

principal_component_pathway_analysis(
  object,
  path = getwd(),
  ppm_error = NULL,
  ion_mode = NULL,
  tof_resolution = 30000,
  num_retained_component = NULL,
  variance_explained_threshold = 0.9,
  resampling_factor = 2,
  p_val_threshold = 0.05,
  byrow = T,
  assay = "SPM",
  slot = "counts",
  flip_plot = FALSE,
  show_variance_plot = FALSE,
  verbose = TRUE
)

Arguments

object

SpaMTP Seurat class object that contains spatial metabolic information.

path

Character string defining the output path for the visualization (default = getwd()).

ppm_error

is the parts-per-million error tolerance of matching m/z value with potential metabolites

ion_mode

is only needed when ppm_error is not specified, used to access the ppm_error based on polarity. Inputs must be either 'positive' or 'negative'(default = NULL).

tof_resolution

is the tof resolution of the instrument used for MALDI run, calculated by ion (ion mass,m/z)/(Full width at half height)

num_retained_component

is an integer value to indicated preferred number of PCs to retain

variance_explained_threshold

Numeric value defining the explained variance threshold (default = 0.9).

resampling_factor

is a numerical value >0, indicate how you want to resample the size of roginal matrix

p_val_threshold

is the p value threshold for pathways to be significant

byrow

is a boolean to indicates whether each column of the matrix is built byrow or bycol.

assay

Character string defining the SpaMTP assay to extract intensity values from (default = "SPM").

slot

Character string defining the assay slot contatin ght intesity values (default = "counts").

flip_plot

Boolean defining whether to rotate the plot 90 degrees (default = FALSE).

show_variance_plot

Boolean indicating weather to display the variance plot output by this analysis (default = FALSE).

verbose

Boolean indicating whether to show the message. If TRUE the message will be show, else the message will be suppressed (default = TRUE).

Value

PCA's and pathway_enrichment_pc is the pathway enrichment results for each PC

Examples

#principal_component_pathway_analysis(mass_matrix = readRDS("~/mass_matrix.rds")[,1:150],width = 912,height = 853,ppm_error = NULL,ion_mode = "positive",tof_resolution = 30000,input_mz = NULL,num_retained_component = NULL,variance_explained_threshold = 0.9,resampling_factor = 2,p_val_threshold = 0.05)