Annotates m/z values based on reference metabolite dataset

Annotates m/z values based on reference metabolite dataset

Usage

AnnotateSM(
  data,
  db,
  feature.metadata.assay = "Spatial",
  feature.metadata.slot = "raw_mz",
  ppm_error = 5,
  test_add_pos = c("M+H"),
  polarity = "pos",
  filepath = NULL,
  return.only.annotated = TRUE,
  verbose = TRUE
)

Arguments

data

Seurat Spatial Metabolomic Object containing m/z values for annotation.

db

Reference metabolite dataset in the form of a Data.Frame.

feature.metadata.assay

Character string defining the Seurat assay which contains the mz counts being annotated (default = "Spatial").

feature.metadata.slot

Character string defining the Seurat assay slot which contains the raw mz values, this is without the 'mz-' and are a vector of integers. This is setup by default when running the cardinal_to_seurat() function (default = "raw_mz").

ppm_error

Numeric value indicating the size of the ppm error allowed when matching molecular weights between Seurat object and reference dataset. If only want exact matches set ppm = 0 (default = 5).

test_add_pos

List of adducts to use for searching the database (e.g. "M+NH4","M+Na","M+CH3OH+H","M+K" etc.) Look at db_adduct_filter() documentation for list of the possible adducts (default = "M+H").

polarity

Character string defining the polarity of adducts to use, either "pos" or "neg" (default = "pos").

filepath

Character string of the directory to store the _annotated_mz_peaks.csv. Else, set to NULL as default.

return.only.annotated

Boolean value indicating if the annotated Seurat Object should only include m/z values that were successfully annotated (default = TRUE).

verbose

Boolean indicating whether to show the message. If TRUE the message will be show, else the message will be suppressed (default = TRUE).

Value

A Seurat Object with m/z values annotated. These annotations are stored in the relative assay's meta.data (e.g. SeuratObj[["Spatial"]][[]])

Examples

# HMDB_db <- load("data/HMDB_1_names.rds")
# Annotated_SeuratObj <- AnnotateSM(SeuratObj, HMDB_db)