Converts and aggregates Spatial Metabolomic (MALDI) data to corresponding Spatial Transcriptomics (Visium) spots.

This function uses generate_new_SM_counts() and increase_SM_res()

Converts and aggregates Spatial Metabolomic (MALDI) data to corresponding Spatial Transcriptomics (Visium) spots.

This function uses generate_new_SM_counts() and increase_SM_res()

MapSpatialOmics(
  SM.data,
  ST.data,
  res_increase = NULL,
  annotations = FALSE,
  assay = "Spatial",
  slots = c("counts"),
  img_res = "hires",
  slice = "slice1",
  new_SpM.assay = "SPM",
  add.ST = TRUE,
  ST.assay = "Spatial",
  ST.layers = c("counts"),
  new_SpT.assay = "SPT",
  verbose = TRUE
)

Arguments

SM.data

A Seurat object representing the Spatial Metabolomics data.

ST.data

A Seurat object representing the Visium Spatial Transcriptomics data.

res_increase

Integer value defining the factor by which the resolution of MALDI spots should be increased before assignment. It should be either 4 or 9, see increase_SM_res() documentation for specifics (Default = NULL).

annotations

Boolean value indicating if the Spatial Metabolomics (MALDI) Seurat object contains annotations assigned to m/z values (default = FALSE).

assay

Character string defining the Seurat assay that contains the annotated counts and metadata corresponding to the m/z values (default = "Spatial").

slots

Vector of character strings describing which slots from the Spatial Metabolomic Seurat object to adjust for the new overlayed object (default = c("counts", "data")).

img_res

Character string defining the image resolution associated with the Visium image pixel data (default = "hires").

slice

Character string of the image slice name in the Visium object (default = "slice1").

new_SpM.assay

Character string defining the assay name of the new overlayed Seurat object containing all updated metabolomic data (default = "SPM").

add.ST

Boolean indicating whether to add the Spatial Metabolomic data from the Seurat Visium Object to the new updated Seruat Object (default = TRUE).

ST.assay

Character string specifying the current assay to use to extract transcriptional data from (default = "Spatial").

ST.layers

Vector of character strings defining the relative slots to extract from the initial Visium object, to add to the new Seurat Object if ST.assay == TRUE (default = c("counts", "data")).

new_SpT.assay

Character string defining the assay name of the new overlayed Seurat object containing all updated transcriptomics data (default = "SPT").

verbose

Boolean value indicating whether to print pregression update messages and progress bar (default = TRUE).

Value

A Seurat object with the Spatial Metabolomic data assigned to equivalent Spatial Transcripomics (Visium) spots.

Examples

## Convert MALDI data to equivalent Visium spots
# MapSpatialOmics(VisiumObj, SeuratObj, img_res = "hires", new_library_id = "MALDI", res_increase = NULL)