Helper function for AnnotatesSM() and FishersPathwayAnalysis()

Helper function for AnnotatesSM() and FishersPathwayAnalysis()

annotateTable(
  mz_df,
  db,
  ppm_error = NULL,
  adducts = NULL,
  polarity = "positive",
  tof_resolution = 30000,
  verbose = TRUE
)

Arguments

mz_df

dataframe containing m/z values for annotation.

db

Reference metabolite dataset in the form of a Data.Frame.

ppm_error

Numeric value indicating the size of the ppm error allowed when matching molecular weights between Seurat object and reference dataset. If only want exact matches set ppm = 0 (default = NULL).

adducts

List of adducts to use for searching the database (e.g. "M+NH4","M+Na","M+CH3OH+H","M+K" etc.). For all possible adducts please visit here. If NULL will take the full list of adducts (default = NULL).

polarity

Character string defining the polarity of adducts to use, either "positive", "negative" or "neutral" (default = "positive").

tof_resolution

is the tof resolution of the instrument used for MALDI run, calculated by ion [ion mass,m/z]/[Full width at half height]. This value is used to estimate ppm_error when set to NULL (default = 30000).

verbose

Boolean indicating whether to show the message. If TRUE the message will be show, else the message will be suppressed (default = TRUE).

Value

Generates an intermediate annotated m/z dataframe

Examples

### HelperFunction